cooperative_demand_response_marl/scripts/evaluate.py
#!/usr/bin/env python3
"""
Script de avaliação para o sistema de resposta cooperativa à demanda
"""

import os
import sys
import yaml
import logging
from pathlib import Path

# Adiciona o diretório src ao path
sys.path.append(str(Path(__file__).parent.parent / "src"))

def setup_logging():
    """Configura o sistema de logging"""
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        handlers=[
            logging.FileHandler('results/logs/evaluation.log'),
            logging.StreamHandler()
        ]
    )
    return logging.getLogger(__name__)

def load_config(config_path):
    """Carrega as configurações do arquivo YAML"""
    with open(config_path, 'r') as file:
        return yaml.safe_load(file)

def main():
    """Função principal de avaliação"""
    logger = setup_logging()
    logger.info("Iniciando sistema de avaliação...")
    
    # Carrega configurações
    config_path = Path(__file__).parent.parent / "config.yaml"
    config = load_config(config_path)
    logger.info(f"Configurações carregadas: {config['project']['name']}")
    
    # TODO: Implementar a avaliação do modelo
    logger.info("Avaliação do modelo a ser implementada...")
    
    logger.info("Avaliação finalizada!")

if __name__ == "__main__":
    main()