cooperative_demand_response_marl/requirements.txt
absl-py==2.3.1
alabaster==1.0.0
babel==2.17.0
beautifulsoup4==4.8.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
CityLearn==2.3.1
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
coverage==7.10.7
cycler==0.12.1
decorator==5.2.1
docutils==0.21.2
doe-xstock==1.1.0
eppy==0.5.63
Farama-Notifications==0.0.4
filelock==3.19.1
flake8==7.3.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
grpcio==1.75.1
gymnasium==1.2.1
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mccabe==0.7.0
mpmath==1.3.0
munch==4.0.0
mypy_extensions==1.1.0
narwhals==2.7.0
networkx==3.5
NREL-PySAM==7.1.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openstudio==3.3.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pettingzoo==1.25.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pluggy==1.6.0
protobuf==6.32.1
pyarrow==21.0.0
pycodestyle==2.14.0
pydot3k==1.0.17
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
roman-numerals-py==3.1.0
scikit-learn==1.2.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stable_baselines3==2.7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tinynumpy==1.2.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
